requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
confluent-kafka==2.3.0
protobuf==6.33.0
typing_extensions==4.15.0
psutil==5.9.8
etcd3==0.12.0
python-dotenv==1.0.1
